little bit of reestructuring

+++ packages/population_repository/lib/src/population_repository.dart
@@ -6,7 +6,7 @@ class PopulationRepository {
   const PopulationRepository();
 
   /// Return the percent of casualities at a given age
-  double calculateCasualities(int age) {
+  int calculateCasualities(int age) {
     return 0;
   }
 }
